@@ -1,21 +1,21 @@
 import React from "react";
 import { useGlobalContext } from "./context";
 const CartItem = ({ id, img, title, price, amount }) => {
-  const { remove, increase, decrease } = useGlobalContext();
+  const { remove, increase, decrease, toggleAmount } = useGlobalContext();
   return (
     <article className="cart-item">
       <img src={img} alt={title} />
       <div>
         <h4>{title}</h4>
-        <h4 className="item-price">PLN {price}</h4>
+        <h4 className="item-price">${price}</h4>
         {/* remove button */}
         <button className="remove-btn" onClick={() => remove(id)}>
-          usuń
+          remove
         </button>
       </div>
       <div>
         {/* increase amount */}
-        <button className="amount-btn" onClick={() => increase(id)}>
+        <button className="amount-btn" onClick={() => toggleAmount(id, "inc")}>
           <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20">
             <path d="M10.707 7.05L10 6.343 4.343 12l1.414 1.414L10 9.172l4.243 4.242L15.657 12z" />
           </svg>
@@ -23,7 +23,7 @@ const CartItem = ({ id, img, title, price, amount }) => {
         {/* amount */}
         <p className="amount">{amount}</p>
         {/* decrease amount */}
-        <button className="amount-btn" onClick={() => decrease(id)}>
+        <button className="amount-btn" onClick={() => toggleAmount(id, "dec")}>
           <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20">
             <path d="M9.293 12.95l.707.707L15.657 8l-1.414-1.414L10 10.828 5.757 6.586 4.343 8z" />
           </svg>
