@@ -1,23 +1,32 @@
-export default [
+const data = [
   {
-    id: 1,
-    title: "Samsung Galaxy S7",
-    price: 599.99,
-    img: "https://res.cloudinary.com/diqqf3eq2/image/upload/v1583368215/phone-2_ohtt5s.png",
+    id: "rec1JZlfCIBOPdcT2",
+    title: "Samsung Galaxy S8",
+    price: "399.99",
+    img: "https://www.course-api.com/images/cart/phone-1.png",
     amount: 1,
   },
   {
-    id: 2,
-    title: "google pixel ",
-    price: 499.99,
-    img: "https://res.cloudinary.com/diqqf3eq2/image/upload/v1583371867/phone-1_gvesln.png",
+    id: "recB6qcHPxb62YJ75",
+    title: "google pixel",
+    price: "499.99",
+    img: "https://www.course-api.com/images/cart/phone-2.png",
     amount: 1,
   },
   {
-    id: 3,
+    id: "recdRxBsE14Rr2VuJ",
     title: "Xiaomi Redmi Note 2",
-    price: 699.99,
-    img: "https://res.cloudinary.com/diqqf3eq2/image/upload/v1583368224/phone-3_h2s6fo.png",
+    price: "699.99",
+    img: "https://www.course-api.com/images/cart/phone-3.png",
+    amount: 1,
+  },
+  {
+    id: "recwTo160XST3PIoW",
+    title: "Samsung Galaxy S7",
+    price: "599.99 ",
+    img: "https://www.course-api.com/images/cart/phone-4.png",
     amount: 1,
   },
 ];
+
+export default data;
