@@ -53,7 +53,22 @@ const reducer = (state, action) => {
   if (action.type === "DISPLAY_ITEMS") {
     return { ...state, cart: action.payload, loading: false };
   }
-
+  if (action.type === "TOGGLE_AMOUNT") {
+    let tempCart = state.cart
+      .map((cartItem) => {
+        if (cartItem.id === action.payload.id) {
+          if (action.payload.type === "inc") {
+            return { ...cartItem, amount: cartItem.amount + 1 };
+          }
+          if (action.payload.type === "dec") {
+            return { ...cartItem, amount: cartItem.amount - 1 };
+          }
+        }
+        return cartItem;
+      })
+      .filter((cartItem) => cartItem.amount !== 0);
+    return { ...state, cart: tempCart };
+  }
   throw new Error("no matching action type");
 };
 
